do not fail if image is not found

+++ script/docker/clean.sh
@@ -8,7 +8,11 @@ SCRIPT_DIRECTORY=$(cd "${DIRECTORY}" || exit 1; pwd)
 script/docker/remove.sh
 
 # Remove image.
-docker rmi "${VENDOR_NAME_LOWER}/${PROJECT_NAME_DASH}"
+docker images | grep --quiet "${PROJECT_NAME_DASH}" && FOUND=true || FOUND=false
+
+if [ "${FOUND}" = true ]; then
+    docker rmi "${VENDOR_NAME_LOWER}/${PROJECT_NAME_DASH}"
+fi
 
 # Remove dangling image identifiers, and more.
 docker system prune
